% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrain.R
\name{dtrain}
\alias{dtrain}
\title{dtrain}
\usage{
dtrain(y)
}
\arguments{
\item{y}{a number or vector}
}
\value{
the height of the density f(x)
}
\description{
This is the density function f for the Train Problem (`d-stem`)
}
\examples{
\dontrun{ dtrain(y = 1.96)}
}
