% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytrain.R
\name{mytrain}
\alias{mytrain}
\title{mytrain}
\usage{
mytrain(a, b)
}
\arguments{
\item{a}{lower endpoint}

\item{b}{upper endpoint such that `P(a <= Y <= b) = P(Y = b) - P(Y = a)`}
}
\value{
a plot with shaded probability area and the probability
}
\description{
This is the precursor to the shiny app included in this package. It produces a plot with shaded region representing the probability between inputs `a` and `b`
}
\examples{
\dontrun{mytrain(a = -2, b = 3)}
}
