% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrain.R
\name{ptrain}
\alias{ptrain}
\title{ptrain}
\usage{
ptrain(y)
}
\arguments{
\item{y}{a number or vector}
}
\value{
the lower-tail probability P(X <= x)
}
\description{
This is the cumulative function F for the Train problem of f (`p-stem`)
}
\examples{
\dontrun{ ptrain(y = 3)}
}
