% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrain.R
\name{rtrain}
\alias{rtrain}
\title{rtrain}
\usage{
rtrain(n, w = 0.3)
}
\arguments{
\item{n}{number of samples}

\item{w}{the value of w where W ~ Unif(0, 1)}
}
\value{
a random sample from the distribution, a histogram of the simulated values, and a plot of the objective function
}
\description{
Produces a random sample from the distribution, a histogram of the simulated values, and a plot of the objective function from `w-F` theory (`r-stem`)
Note also that using a specific value for w yields the "p-train" function (`p-stem`)
}
\examples{
\dontrun{rtrain(n = 10000, w = 0.3)}
}
