% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytrains.R
\name{shinytrains}
\alias{shinytrains}
\title{shinytrains}
\usage{
shinytrains()
}
\value{
a plot of the density with shaded area and displayed probability result
}
\description{
An interactive shiny app which allows the user to engage a slider to change the values for which we'd like to know the probability between (area under) the plotted density curve. Further the app features the ability to change the values by "brushing".
}
\examples{
\dontrun{shinytrains()}
}
